Tabel werkt Data wordt aangepast in FestivalData, hierbij is een SmallpositiveintegerField veranderd naar een FloatField om errors te voorkomen. Mogelijk moeten nog een aantal dingen worden ingebouwd om nog meer errors te voorkomen.

+++ hielander_whiskey_app/models/FestivalData.py
@@ -23,4 +23,4 @@ class FestivalData(models.Model):
     )
     aantal_beschikbaar = models.PositiveSmallIntegerField()
 
-    prijs = models.PositiveSmallIntegerField()
+    prijs = models.FloatField()
